Using deploy target in artifact capturing scripts

+++ hack/capture_artifacts.sh
@@ -1,29 +1,25 @@
     #!/bin/bash -x
     echo "Capturing the System Information"
-    if [ -d ${WORKSPACE}/canary-deployments ];then
-        cd ${WORKSPACE}/canary-deployments
+    if [ -d ${WORKSPACE}/deploy ];then
+        cd ${WORKSPACE}/deploy
     else
         exit 1 
     fi
     # Capturing Teraform template
-    if [ ! -f ${WORKSPACE}/canary-deployments/.deploy-power-powervc.tfvars.template ]; then
-        echo "${WORKSPACE}/canary-deployments/.deploy-power-powervc.tfvars.template not found!"
+    if [ ! -f ${WORKSPACE}/deploy/.${TARGET}.tfvars ]; then
+        echo "${WORKSPACE}/deploy/.${TARGET}.tfvars not found!"
         exit 1
     else
-        cp ${WORKSPACE}/canary-deployments/.deploy-power-powervc.tfvars.template deploy-power-powervc.tfvars.template
-        sed -i "s|ftp3_password.*$|ftp3_password  = ************|g" deploy-power-powervc.tfvars.template
+        cp ${WORKSPACE}/deploy/.${TARGET}.tfvars ${TARGET}.tfvars
+        sed -i "s|ftp3_password.*$|ftp3_password  = ************|g" ${TARGET}.tfvars
+        sed -i "s|repo_token.*$|repo_token  = ************|g" ${TARGET}.tfvars
+        sed -i "s|icp_version.*=.*@|icp_version = \"user:************@|g" ${TARGET}.tfvars
+        sed -i "s|docker_password.*$|docker_password: ************|g" ${TARGET}.tfvars
+        sed -i "s|header.*X-JFrog-Art-Api.*$|header.*X-JFrog-Art-Api: ************|g" ${TARGET}.tfvars
+        sed -i "s|password.*=.*$|password = ************|g" ${TARGET}.tfvars
+        sed -i "s|rhel_subscription_password.*=.*$|prhel_subscription_password = ************|g" ${TARGET}.tfvars
     fi
-    if [ ! -f .deploy-power-powervc.tfvars ]; then
-        echo ".deploy-power-powervc.tfvars not found!"
-        exit 1
-    else
-        cp .deploy-power-powervc.tfvars deploy-power-powervc.tfvars
-        sed -i "s|ftp3_password.*$|ftp3_password  = ************|g" deploy-power-powervc.tfvars
-        sed -i "s|icp_version.*=.*@|icp_version = \"user:************@|g" deploy-power-powervc.tfvars
-        sed -i "s|docker_password.*$|docker_password: ************|g" deploy-power-powervc.tfvars
-        sed -i "s|header.*X-JFrog-Art-Api.*$|header.*X-JFrog-Art-Api: ************|g" deploy-power-powervc.tfvars
-    fi
-    MASTER_NODE=$(make terraform:output TERRAFORM_DIR=.deploy-power-powervc TERRAFORM_OUTPUT_VAR=master-node || true)
+    MASTER_NODE=$(make terraform:output TERRAFORM_DIR=.${TARGET} TERRAFORM_OUTPUT_VAR=master-node || true)
     [ $? -ne 0 ] && exit 1;
     if [ ! -z "${MASTER_NODE}" ]; then
         ssh -q -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} exit
@@ -32,14 +28,17 @@
             scp -i id_rsa -o StrictHostKeyChecking=no ${WORKSPACE}/${PROJECTNAME}/hack/system_info.sh root@${MASTER_NODE}:
             ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} /root/system_info.sh > systeminfo.txt
             tar -czvf systeminfo.txt.tar.gz systeminfo.txt
-            if ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} '[ -d /opt/ibm/cluster ]' ;then
-                scp -i id_rsa -o StrictHostKeyChecking=no ${WORKSPACE}/${PROJECTNAME}/hack/health_check.sh root@${MASTER_NODE}:/opt/ibm/cluster/
-                ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} rm -rf /opt/ibm/cluster/images
-                ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} /opt/ibm/cluster/health_check.sh
-                scp -i id_rsa -o StrictHostKeyChecking=no -r root@${MASTER_NODE}:/opt/ibm/cluster .
-                sed -i "s|header: 'X-JFrog-Art-Api.*$|header: 'X-JFrog-Art-Api: ************|g" cluster/config.yaml
-                sed -i "s|docker_password.*$|docker_password: ************|g" cluster/config.yaml
-                tar -czvf cluster.tar.gz cluster
+            if [ ${SKIP_ICP_INSTALL} != "true" ]; then
+                if ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} '[ -d /opt/ibm/cluster ]' ;then
+                    scp -i id_rsa -o StrictHostKeyChecking=no ${WORKSPACE}/${PROJECTNAME}/hack/health_check.sh root@${MASTER_NODE}:/opt/ibm/cluster/
+                    ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} rm -rf /opt/ibm/cluster/images
+                    ssh -i id_rsa -o StrictHostKeyChecking=no root@${MASTER_NODE} /opt/ibm/cluster/health_check.sh
+                    scp -i id_rsa -o StrictHostKeyChecking=no -r root@${MASTER_NODE}:/opt/ibm/cluster .
+                    sed -i "s|header: 'X-JFrog-Art-Api.*$|header: 'X-JFrog-Art-Api: ************|g" cluster/config.yaml
+                    sed -i "s|docker_password.*$|docker_password: ************|g" cluster/config.yaml
+                    sed -i "s|rhel_subscription_password.*$|rhel_subscription_password: ************|g" cluster/config.yaml
+                    tar -czvf cluster.tar.gz cluster
+                fi
             fi
         else
             echo 'Unable to access the cluster. You may delete the VMs manually'
